Add configured backend and seperated electrum functions
